Resolving build errors.

+++ src/CEDS-Data-Warehouse-Project/RDS/Tables/BridgeK12AcademicCalendarGradelevels.sql
@@ -1,6 +1,6 @@
 CREATE TABLE [RDS].[BridgeK12AcademicCalendarGradeLevels] (
     [BridgeK12AcademicCalendarGradeLevelId] INT IDENTITY (1, 1) NOT NULL,
-    [FactK12AcademicCalendarId]           INT NOT NULL,
+    [FactK12AcademicCalendarId]           BIGINT NOT NULL,
     [GradeLevelId]                              INT NOT NULL,
     CONSTRAINT [PK_BridgeK12AcademicCalendarGradeLevels] PRIMARY KEY CLUSTERED ([BridgeK12AcademicCalendarGradeLevelId] ASC),
     CONSTRAINT [FK_BridgeK12AcademicCalendarGradeLevels_FactK12AcademicCalendarId] FOREIGN KEY ([FactK12AcademicCalendarId]) REFERENCES [RDS].[FactK12AcademicCalendars] ([FactK12AcademicCalendarId]),
